requirements.txt
arrow==1.2.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==1.7.3
Django==4.0.2
django-browser-reload==1.3.0
django-tailwind==3.1.1
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.1.0
numpy==1.22.2
Pillow==9.0.1
poyo==0.5.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.0.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
whitenoise==6.0.0
